Option for meteor settings file
Some applications use a settings file, given via the --settings argument to the
meteor/mrt command or as a JSON string assigned to the METEOR_SETTINGS
environmental variable on self-hosted environments.

By providing an argument to --settings it is now possible to provide the
location of the meteor settings file and this is read, stringified and assigned
to the METEOR_SETTINGS environmental variable for each app.